third_party/WebKit/LayoutTests/external/wpt/css-timing-1/frames-timing-functions-output-expected.txt

This is a testharness.js-based test.
FAIL For an input progress of 0.0, the output of a frames timing function is the first frame assert_equals: expected "0px" but got "auto"
FAIL At a frame boundary, the output of a frames timing function is the next frame assert_equals: expected "0px" but got "auto"
FAIL For an input progress of 1.0, the output of a frames timing function is the final frame assert_equals: expected "100px" but got "auto"
FAIL The number of frames is correctly reflected in the frames timing function output assert_equals: expected "0px" but got "auto"
FAIL The number of frames is correctly reflected in the frames timing function output on CSS Transitions assert_equals: expected "0px" but got "100px"
FAIL frames easing with input progress greater than 1 Failed to execute 'animate' on 'Element': 'frames(2)' is not a valid value for easing
FAIL frames easing with input progress greater than 1.5 Failed to execute 'animate' on 'Element': 'frames(2)' is not a valid value for easing
FAIL frames easing with input progress less than 0 Failed to execute 'animate' on 'Element': 'frames(2)' is not a valid value for easing
Harness: the test ran to completion.
